42 -> GH 20221117[1]

+++ ft_strjoin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: antoda-s <antoda-s@student.42porto.com>    +#+  +:+       +#+        */
+/*   By: antoda-s <antoda-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 00:28:11 by antoda-s          #+#    #+#             */
-/*   Updated: 2022/11/14 00:55:05 by antoda-s         ###   ########.fr       */
+/*   Updated: 2022/11/17 13:48:24 by antoda-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned int len;
-	char *dest;
-	
+	size_t	len;
+	char	*dest;
+
 	if (!s1 || !s2)
 		return (NULL);
 	len = ft_strlen(s1)+ft_strlen(s2)+1;
-    dest = (char*)malloc(len);
+	dest = (char *)malloc(len);
 	if (!dest)
 		return (NULL);
 	while (*s1)
@@ -31,7 +31,7 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	{
 		*dest++ = *s2++;
 	}
-    *dest = '\0';
+	*dest = '\0';
 	dest -= (len - 1);
-    return (dest);
+	return (dest);
 }
